requirements.txt
django==1.11
Pillow
djangorestframework